requirements.txt
numpy
black
ipython
absl-py
